refactor(constants): 重構常量類別，將 Constants 類別替換為 Constant 類別
